refactor: GameResult 이름 변경, Player 결과 산청 로직 when으로 변경
